adding animation forward/backward and transformation flip, sequence clone and renaming. also mouse interaction(rotation and wheel based distance) on the model preview. billboards are not created if and sceneobject is hidden. GetSceneObjectsByType will not return objects marked for deletion. SceneObjectsStep and GameStep call was swaped to fix inconsistent writes to Renderables which are not picked by the steps functions like AnimationSteps

+++ Engine/Source/pch/NoMath.h
@@ -155,12 +155,12 @@ inline XMFLOAT3 Quaternion2Euler(XMVECTOR Q)
 	float test = x * y + z * w;
 	if (test > 0.499 * unit) { // singularity at north pole
 		angles.y = 2 * std::atan2(x, w);
-		angles.z = M_PI_2;
+		angles.z = static_cast<float>(M_PI_2);
 		angles.x = 0;
 	}
 	else if (test < -0.499 * unit) { // singularity at south pole
 		angles.y = -2 * std::atan2(x, w);
-		angles.z = -M_PI_2;
+		angles.z = -static_cast<float>(M_PI_2);
 		angles.x = 0;
 	}
 	else {
